#11 - refactor(payment/wallet): WalletController·Repository 패키지 이동 및 조회 메서드 정리
- WalletController, WalletRepository 를 payment.wallet 패키지로 이동
  - payment 도메인 내 하위 모듈 구조 정리
- Controller: findWalletByUserId → findWalletByWalletId 로 호출 대상 변경
  - 서비스/도메인 모델의 지갑 조회 정책과 일관성 확보
- Repository: 기존 로직 유지하되 위치 변경
